// source/ucp-infra/custom_resource/lib/enums.ts
// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
// SPDX-License-Identifier: Apache-2.0

export enum CustomResourceActions {
    CREATE_UUID = 'CreateUuid',
    DEPLOY_FRONT_END = 'DeployFrontEnd'
}

export enum CustomResourceRequestTypes {
    CREATE = 'Create',
    UPDATE = 'Update',
    DELETE = 'Delete'
}

export enum StatusTypes {
    SUCCESS = 'SUCCESS',
    FAILED = 'FAILED'
}